fix: cast `melg64::melg_base::Lag1Over to `std::size_t` for comparison with `melg64::melg_base::i_` in `melg64::melg_base::next_case2`

+++ include/melg64/melg64.hpp
@@ -292,7 +292,7 @@ class melg_base {
 
     ++this->i_;
 
-    if (this->i_ == this->Lag1Over) {
+    if (this->i_ == static_cast<std::size_t>(this->Lag1Over)) {
       this->next_ = &this->next_case3;
     }
